FIX: Also use namespaced templates; with priority.
lest we be doomed to fail on the mordern SS4 system. Templates
are also namespaced, and are encouraged to follow the same pattern
in folder structure as their namespace. We can use the array notation
to fall back to the un-namespaced classname should a developer still
be retaining the old style strucutre, so we get the best of both worlds :)